files refactor; screen layout fixes
